fix: handle async replication changed during initialization
