+++ adapters/repos/db/migrator_shard_status_ops.go
@@ -131,7 +131,7 @@ func (m *Migrator) freeze(ctx context.Context, idx *Index, class string, freeze 
 			defer idx.shardCreateLocks.Unlock(name)
 
 			if shard != nil {
-				if err := shard.HaltForTransfer(ctx); err != nil {
+				if err := shard.HaltForTransfer(true, ctx); err != nil {
 					m.logger.WithFields(logrus.Fields{
 						"action": "halt_for_transfer",
 						"error":  err,
