+++ adapters/repos/db/shard_drop.go
@@ -42,8 +42,7 @@ func (s *Shard) drop() (err error) {
 		s.clearDimensionMetrics()
 	}
 
-	s.mayStopHashBeater()
-	s.mayCleanupHashTree()
+	s.mayStopAsyncReplication()
 
 	ctx, cancel := context.WithTimeout(context.TODO(), 20*time.Second)
 	defer cancel()
